api_v1/serializers.py
from rest_framework import serializers
from api_v1.models import Customer
from front_end.models import UserInfo

class CustomerSerializer(serializers.ModelSerializer):

    class Meta:
        model = Customer
        fields = '__all__'

class UserInfoSerializer(serializers.ModelSerializer):

    class Meta:
        model = UserInfo
        fields = '__all__'
