requirements.txt
Django==2.1.3
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-filter==2.0.0
djangorestframework==3.9.0
gunicorn==19.9.0
Markdown==3.0.1
pytz==2018.7
